requirements.txt
w3lib==2.1.1
requests==2.28.1
beautifulsoup4==4.11.1
selenium==4.6.1
MySQLdb==1.4.6
PyYAML==5.4.1
python-twitter==3.5
